update-ipfs.sh: Add another IPFS node

+++ update-ipfs.sh
@@ -1,11 +1,22 @@
 #!/bin/sh
+#
+# ipfs daemon should be running on the machine
+# where this script is run
 
 DNS=pr.bublina.eu.org
-OLDREF=$(./dnslink.sh $DNS)
+
+updateremote() {
+  ssh $1 ipfs pin add $2
+  ssh $1 ipfs pin rm $3
+}
+
+OLDREF=$(./dnslink.sh $DNS | cut -d/ -f3-)
 NEWREF=$(ipfs add -Qr public)
+test "$OLDREF" = "$NEWREF" && { echo "No change"; exit 1; }
+
 ipfs pin rm $OLDREF
-ssh random ipfs pin add $NEWREF
-ssh random2 ipfs pin add $NEWREF
-ssh random ipfs pin rm $OLDREF
-ssh random2 ipfs pin rm $OLDREF
+for i in $(seq 3)
+do
+  updateremote random$i $NEWREF $OLDREF
+done
 ./dnslink-set.sh $DNS $NEWREF
